factor server build state out to startup module
